refactor: use relative imports inside the integration package
'from custom_components.ucams.X import …' breaks the moment Home
Assistant re-roots the package or pytest is invoked with a different
CWD, and it forces every consumer to round-trip through __init__.py
even when they don't need to. Switch to 'from .X import …' throughout
and pull UcamsApi/DomApi/CONF_* directly from their defining modules
rather than from the package facade.

Drive-bys turned up by ruff after the import cleanup:
- update_link's 'comment: str = None' becomes 'str | None = None'.
- UcamsApi's forward-ref 'DomApi' annotation is replaced with a real
  import (no cycle — ufanet.py doesn't import ucams.py).
